Revert "Add React frontend pages for core FBPA entities"
