+++ src/App.jsx
@@ -2,10 +2,6 @@ import React from 'react';
 import './index.css';
 import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
 import Dashboard from './pages/Dashboard.jsx';
-import Customers from './pages/Customers.jsx';
-import Carriers from './pages/Carriers.jsx';
-import Invoices from './pages/Invoices.jsx';
-import Exceptions from './pages/Exceptions.jsx';
 import LoginForm from './components/LoginForm.jsx';
 import RegisterForm from './components/RegisterForm.jsx';
 
@@ -25,38 +21,6 @@ function App() {
           element={
             <PrivateRoute>
               <Dashboard />
-            </PrivateRoute>
-          }
-        />
-        <Route
-          path="/customers"
-          element={
-            <PrivateRoute>
-              <Customers />
-            </PrivateRoute>
-          }
-        />
-        <Route
-          path="/carriers"
-          element={
-            <PrivateRoute>
-              <Carriers />
-            </PrivateRoute>
-          }
-        />
-        <Route
-          path="/invoices"
-          element={
-            <PrivateRoute>
-              <Invoices />
-            </PrivateRoute>
-          }
-        />
-        <Route
-          path="/exceptions"
-          element={
-            <PrivateRoute>
-              <Exceptions />
             </PrivateRoute>
           }
         />
